FM-114: Move fvm<->evm conversion to the message crate
