+++ fendermint/rpc/Cargo.toml
@@ -24,7 +24,7 @@ fvm_ipld_encoding = { workspace = true }
 fvm_shared = { workspace = true }
 
 fendermint_vm_actor_interface = { path = "../vm/actor_interface" }
-fendermint_vm_message = { path = "../vm/message", features = ["secp256k1"] }
+fendermint_vm_message = { path = "../vm/message" }
 
 [dev-dependencies]
 clap = { workspace = true }
