+++ fendermint/vm/message/Cargo.toml
@@ -7,6 +7,11 @@ edition.workspace = true
 license.workspace = true
 
 [dependencies]
+anyhow = { workspace = true }
+blake2b_simd = { workspace = true }
+ethers-core = { workspace = true }
+lazy_static = { workspace = true }
+libsecp256k1 = { workspace = true }
 thiserror = { workspace = true }
 serde = { workspace = true }
 serde_tuple = { workspace = true }
@@ -17,14 +22,13 @@ cid = { workspace = true }
 fvm_shared = { workspace = true }
 fvm_ipld_encoding = { workspace = true }
 
-libsecp256k1 = { workspace = true, optional = true }
-blake2b_simd = { workspace = true, optional = true }
 
 arbitrary = { workspace = true, optional = true }
 quickcheck = { workspace = true, optional = true }
 rand = { workspace = true, optional = true }
 
 fendermint_vm_encoding = { path = "../encoding" }
+fendermint_vm_actor_interface = { path = "../actor_interface" }
 fendermint_testing = { path = "../../testing", optional = true }
 
 [dev-dependencies]
@@ -36,10 +40,8 @@ hex = { workspace = true }
 # Ideally we could do this with `#[cfg(any(test, feature = "arb"))]`,
 # however in that case all the extra dependencies would not kick in,
 # and we'd have to repeat all those dependencies.
-fendermint_vm_message = { path = ".", features = ["arb", "secp256k1"] }
+fendermint_vm_message = { path = ".", features = ["arb"] }
 fendermint_testing = { path = "../../testing", features = ["golden"] }
 
 [features]
-default = ["secp256k1"]
 arb = ["arbitrary", "quickcheck", "fvm_shared/arb", "cid/arb", "rand", "fendermint_testing/arb"]
-secp256k1 = ["libsecp256k1", "blake2b_simd"]
